+++ src/Web/Masa.Auth.Web.Admin.Rcl/Pages/Sso/CustomLoginRegister/LoginRegisterTemplate.razor.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) MASA Stack All rights reserved.
 // Licensed under the Apache License. See LICENSE.txt in the project root for license information.
 
-namespace Masa.Auth.Web.Admin.Rcl.Pages.Sso.CustomLogin;
+namespace Masa.Auth.Web.Admin.Rcl.Pages.Sso.CustomLoginRegister;
 
 public partial class LoginRegisterTemplate
 {
