+++ src/Web/Masa.Auth.Web.Admin.Rcl/_Imports.razor
@@ -18,4 +18,4 @@
 @using Masa.Stack.Components.GlobalNavigation;
 @using Masa.Blazor.Presets;
 @using Masa.Auth.Web.Admin.Rcl.Pages.Subjects.Teams;
-@using Masa.Auth.Web.Admin.Rcl.Pages.Sso.CustomLogin.Model;
+@using Masa.Auth.Web.Admin.Rcl.Pages.Sso.CustomLoginRegister.Model;
